third_party/blink/web_tests/wpt_internal/prerender/about-blank-iframes.html

<!doctype html>
<!--
Tests for about:blank iframes `document.prerendering` state
and prerenderingchange event.

This file cannot be upstreamed to WPT until:
* startPrerendering() usage is replaced with a WebDriver API.
-->
<title>Test about:blank iframes prerendering state</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
promise_test(async t => {
  const channel = new BroadcastChannel('test-channel');

  const gotMessage = new Promise(resolve => {
    channel.addEventListener('message', e => {
      resolve(e.data);
    }, {once: true});
  });

  // Make the window to start the prerender.
  const url = `resources/about-blank-iframes.html`;
  window.open(url, '_blank', 'noopener');

  const msg = await gotMessage;
  assert_equals(msg, 'PASS');
}, 'about:blank iframes');
</script>
</body>
